Package structure refactor
- Code from template project removed completely
- Package structure is refactored. The packages will be structured by features. (https://medium.com/hackernoon/package-by-features-not-layers-2d076df1964d#.f7znkie19)
