+++ Application/app/src/main/java/hu/bme/aut/android/together/features/search/EventQueryFragment.kt
@@ -1,0 +1,23 @@
+package hu.bme.aut.android.together.features.search
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import hu.bme.aut.android.together.R
+
+class EventQueryFragment : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_event_query, container, false)
+    }
+}
