--- Application/app/src/main/java/hu/bme/aut/android/together/ui/dashboard/DashboardViewModel.kt
@@ -1,13 +1,0 @@
-package hu.bme.aut.android.together.ui.dashboard
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class DashboardViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
-    }
-    val text: LiveData<String> = _text
-}
