+++ Application/app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,18 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
+    <!-- TODO: the drawable icons used by the template should be deleted -->
+
     <item
-        android:id="@+id/navigation_profile"
+        android:id="@+id/profileFragment"
         android:icon="@drawable/ic_action_person"
         android:title="@string/profile" />
 
     <item
-        android:id="@+id/navigation_myEvents"
+        android:id="@+id/currentEventsListFragment"
         android:icon="@drawable/ic_action_event"
         android:title="@string/my_events" />
 
     <item
-        android:id="@+id/navigation_search"
+        android:id="@+id/eventQueryFragment"
         android:icon="@drawable/ic_action_search"
         android:title="@string/search_events" />
