+++ Application/app/src/main/res/values/strings.xml
@@ -3,4 +3,6 @@
     <string name="profile">Profile</string>
     <string name="my_events">My events</string>
     <string name="search_events">Search events</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
